fixed the logic of calculating the inventory after adding it to the cart.
