+++ src/shared/lib/shopify/fragments/index.ts
@@ -1,0 +1,7 @@
+// Export all GraphQL fragments
+export { cartFragment } from "./cart";
+export { imageFragment } from "./image";
+export { optionFragment } from "./option";
+export { productFragment } from "./product";
+export { seoFragment } from "./seo";
+export { variantFragment } from "./variant";
