+++ src/shared/lib/shopify/fragments/variant.ts
@@ -1,0 +1,20 @@
+export const variantFragment = /* GraphQL */ `
+	fragment variant on ProductVariant {
+		id
+		title
+		availableForSale
+		quantityAvailable
+		selectedOptions {
+			name
+			value
+		}
+		price {
+			amount
+			currencyCode
+		}
+		compareAtPrice {
+			amount
+			currencyCode
+		}
+	}
+`;
